change packing method for WeightsVector weight
git-svn-id: http://svn.cern.ch/guest/lhcb/LHCb/trunk@86482 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: ccbace5241a6393fc3b6fed2b43cb7cef7e1ea64

+++ LHCb/Event/PackedEvent/Event/PackedWeightsVector.h
@@ -34,8 +34,8 @@ namespace LHCb
       : key(0), weight(0)
     {}
 
-    int key;           ///< key of the track this weight is associated with
-    short int weight;  ///< Weight of this track in the PV
+    int key;     ///< key of the track this weight is associated with
+    int weight;  ///< Weight of this track in the PV
 
   };
